+++ src/dotnet/Lucene.Net.CodeAnalysis.VisualBasic/Lucene1000_SealIncrementTokenMethodVBCodeFixProvider.cs
@@ -34,7 +34,7 @@ namespace Lucene.Net.CodeAnalysis
     {
         private const string Title = "Add NotOverridable keyword to IncrementToken() method";
 
-        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(Lucene1000_TokenStreamOrItsIncrementTokenMethodMustBeSealedAnalyzer.DiagnosticId);
+        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(Lucene1000_TokenStreamOrItsIncrementTokenMethodMustBeSealedVBAnalyzer.DiagnosticId);
 
         public sealed override FixAllProvider GetFixAllProvider()
         {
